refactor: logger 헤더에서 소스 분리
